ROI documents now are not deleted if the new org is the same as the old

+++ trade_remedies_api/cases/services/v2/views.py
@@ -124,27 +124,30 @@ class SubmissionViewSet(BaseModelViewSet):
         )
 
         # Removing the previous organisation from the case if they are not properly enrolled
-        OrganisationCaseRole.objects.filter(
-            organisation=previous_organisation_object,
-            case=submission_object.case,
-            role=CaseRole.objects.get(id=ROLE_PREPARING),
-        ).delete()
+        if organisation_object != previous_organisation_object:
+            OrganisationCaseRole.objects.filter(
+                organisation=previous_organisation_object,
+                case=submission_object.case,
+                role=CaseRole.objects.get(id=ROLE_PREPARING),
+            ).delete()
 
-        # Associating the organisation with the case
-        OrganisationCaseRole.objects.get_or_create(
-            organisation=organisation_object,
-            case=submission_object.case,
-            defaults={
-                "role": CaseRole.objects.get(id=ROLE_PREPARING),
-                "sampled": True,
-                "created_by": request.user,
-            },
-        )
+            # Associating the organisation with the case
+            OrganisationCaseRole.objects.get_or_create(
+                organisation=organisation_object,
+                case=submission_object.case,
+                defaults={
+                    "role": CaseRole.objects.get(id=ROLE_PREPARING),
+                    "sampled": True,
+                    "created_by": request.user,
+                },
+            )
 
         # Deleting all the user SubmissionDocument objects as they no longer apply to the submission
-        submission_object.submissiondocument_set.filter(
-            type__key__in=["respondent", "loa"]
-        ).delete()
+        # only if the new organisation is different from the previous
+        if organisation_object != previous_organisation_object:
+            submission_object.submissiondocument_set.filter(
+                type__key__in=["respondent", "loa"]
+            ).delete()
 
         submission_object.organisation = organisation_object
         submission_object.contact = contact_object
